Improving some basic stuff
Allowing the enemy to know they were hit by the enemy. Also making sure that the director AI is less likely to get a null pointer exception when an AI has been killed but the array containing all the current enemies in the amount doesn't return null when going through the array.