More abstraction of the states
As I was reworking on the new states so they could in a HFSM I realized that many of the functions that could be placed one level above could also work if it was in the common methods script. Have been slowly moving methods and parts there to more easily control what is being called.